refactor : getImgSrc 유틸함수 사용
