+++ src/components/common/Map.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect, useRef, SetStateAction, Dispatch } from 'react';
-
+import { getSelectedImg, getDefaultImg } from '@utils/getImgSrc';
 type MapProps = {
   shopInfo: Shop[] | undefined;
   isLoaded: boolean;
@@ -21,32 +21,6 @@ const Map = ({
 }: MapProps) => {
   const mapContainer = useRef<HTMLDivElement>(null);
   const [markers, setMarkers] = useState<any[]>([]);
-
-  const getSelectedImg = (name) => {
-    const src = name.includes('인생네컷')
-      ? '/svg/marker_select_pink.svg'
-      : name.includes('하루필름')
-      ? '/svg/marker_select_gray.svg'
-      : name.includes('포토이즘')
-      ? '/svg/marker_select_blue.svg'
-      : name.includes('포토그레이')
-      ? '/svg/marker_select_black.svg'
-      : '/svg/marker_select_white.svg';
-    return src;
-  };
-
-  const getDefaultImg = (name) => {
-    const src = name.includes('인생네컷')
-      ? '/svg/marker_pink.svg'
-      : name.includes('하루필름')
-      ? '/svg/marker_gray.svg'
-      : name.includes('포토이즘')
-      ? '/svg/marker_blue.svg'
-      : name.includes('포토그레이')
-      ? '/svg/marker_black.svg'
-      : '/svg/marker_white.svg';
-    return src;
-  };
 
   useEffect(() => {
     if (isLoaded) {
